requirements.txt
Flask==3.1.3
gunicorn==25.1.0
flask-sqlalchemy==3.1.1
pymysql==1.1.2
python-dotenv==1.2.2
cryptography==46.0.5
Flask-Migrate==4.1.0
argon2-cffi==25.1.0
flask-talisman==1.1.0
flask-wtf==1.2.2
flask-session==0.8.0
redis==7.3.0
flask-limiter==4.1.1
Pillow==12.1.1
PyMuPDF==1.27.2